django_etl/sofa_sport/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
